% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{academie}
\alias{academie}
\title{Table academie}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 34 rows and 2 columns.
}
\usage{
academie
}
\description{
\preformatted{## Rows: 34
## Columns: 2
## $ code_academie <chr> "25", "20", "16", "23", "02", "13", "08", "27", "32", "40", "22", "03", "...
## $ lib_academie  <chr> "Versailles", "Amiens", "Toulouse", "Nice", "Aix-Marseille", "Poitiers", ...
}
}
\keyword{datasets}
