% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libelle.R
\name{acronyme_bac}
\alias{acronyme_bac}
\title{Renvoie l'acronyme a partir du code Bac}
\usage{
acronyme_bac(code_bac)
}
\arguments{
\item{code_bac}{Un vecteur de code Bac.}
}
\value{
Un vecteur contenant les acronymes de Bac.

Jeu de données source : \code{apogee::bac}.\cr
Il est créé à partir de la table "bac" de la base Access "Tables_ref.accdb" (projet Apogée).
}
\description{
Renvoie l'acronyme à partir du code Bac.
}
