% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libelle.R
\name{acronyme_domaine_diplome}
\alias{acronyme_domaine_diplome}
\title{Renvoie l'acronyme a partir du code de domaine de diplome}
\usage{
acronyme_domaine_diplome(code_domaine_diplome)
}
\arguments{
\item{code_domaine_diplome}{Un vecteur de code de domaine de diplôme.}
}
\value{
Un vecteur contenant les acronymes de domaine de diplôme.

Jeu de données source : \code{apogee::domaine_diplome}.\cr
Il est créé à partir de la table "apogee::domaine_diplome" de la base Access "Tables_ref.accdb".
}
\description{
Renvoie l'acronyme à partir du code de domaine de diplôme.
}
