% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libelle.R
\name{acronyme_etape}
\alias{acronyme_etape}
\title{Renvoie l'acronyme a partir du code etape}
\usage{
acronyme_etape(code_etape, prefixe = "formation", suffixe = c("ville",
  "option", "particularite", "annee"))
}
\arguments{
\item{code_etape}{Un vecteur de code étape.}

\item{prefixe}{Libellé de type de formation ou de type de diplôme en préfixe du libellé de formation.}

\item{suffixe}{Libellé de ville, option, particularité ou année dans le diplôme en suffixe de l'acronyme de formation.}
}
\value{
Un vecteur contenant les acronymes d'étape.

Jeu de données source : \code{apogee::etape}.\cr
Il est créé à partir de la table "etape" de la base Access "Tables_ref.accdb".
}
\description{
Renvoie l'acronyme à partir du code étape.
}
