% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libelle.R
\name{acronyme_mention_diplome}
\alias{acronyme_mention_diplome}
\title{Renvoie l'acronyme a partir du code de mention de diplome}
\usage{
acronyme_mention_diplome(code_mention_diplome)
}
\arguments{
\item{code_etape}{Un vecteur de code de mention de diplôme.}
}
\value{
Un vecteur contenant les acronymes de mention de diplôme.

Jeu de données source : \code{apogee::mention_diplome}.\cr
Il est créé à partir de la table "mention_diplome" de la base Access "Tables_ref.accdb".
}
\description{
Renvoie l'acronyme à partir du code de mention de diplôme.
}
