% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libelle.R
\name{acronyme_type_diplome}
\alias{acronyme_type_diplome}
\title{Renvoie l'acronyme a partir du code de type diplome}
\usage{
acronyme_type_diplome(code_type_diplome)
}
\arguments{
\item{code_type_diplome}{Un vecteur de code de type diplôme.}
}
\value{
Un vecteur contenant les acronymes de type diplôme.

Jeu de données source : \code{apogee::diplome_type}.\cr
Il est créé à partir de la table "diplome_type" de la base Access "Tables_ref.accdb" (projet Apogée).
}
\description{
Renvoie l'acronyme à partir du code de type diplôme.
}
