% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{annee_en_cours}
\alias{annee_en_cours}
\title{Renvoie l'annee universitaire en cours}
\usage{
annee_en_cours()
}
\value{
Une valeur integer contenant l'année en cours
}
\description{
Renvoie l'année universitaire en cours selon la dernière année observée dans Apogée.
}
