% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bac}
\alias{bac}
\title{Table bac}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 139 rows and 4 columns.
}
\usage{
bac
}
\description{
\preformatted{## Rows: 139
## Columns: 4
## $ code_bac        <chr> "S4", "0000", "0001", "0002", "0021", "0022", "0023", "0030", "0031", "...
## $ lib_bac         <chr> "Equivalence", "0000-sans", "0001-intl", "0002-europ", "0021-indus", "0...
## $ acronyme_bac    <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA,...
## $ code_bac_parent <chr> NA, "01", "01", "01", "09", "09", "09", "01", "01", "01", "01", "01", "...
}
}
\keyword{datasets}
