% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bac_mention}
\alias{bac_mention}
\title{Table bac_mention}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 4 rows and 2 columns.
}
\usage{
bac_mention
}
\description{
\preformatted{## Rows: 4
## Columns: 2
## $ code_mention_bac <chr> "TB", "B", "AB", "P"
## $ lib_mention_bac  <chr> "Très bien", "Bien", "Assez bien", "Passable"
}
}
\keyword{datasets}
