% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bce_lycees}
\alias{bce_lycees}
\title{Table bce_lycees}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 5297 rows and 17 columns.
}
\usage{
bce_lycees
}
\description{
\preformatted{## Rows: 5,297
## Columns: 17
## $ numero_uai                <chr> "0010001W", "0010006B", "0010010F", "0010013J", "0010014K", "...
## $ appellation_officielle    <chr> "Lycée professionnel Alexandre Bérard", "Lycée polyvalent Sai...
## $ denomination_principale   <chr> "LP LYCEE DES METIERS", "LPO LYCEE DES METIERS", "LYCEE GENER...
## $ patronyme_uai             <chr> "ALEXANDRE BERARD", "SAINT-EXUPERY", "DU BUGEY", "LALANDE", "...
## $ secteur_public_prive_libe <chr> "Public", "Public", "Public", "Public", "Public", "Public", "...
## $ adresse_uai               <chr> "223  RUE ALEXANDRE BERARD", "15  AV SAINT-EXUPERY", "113 RUE...
## $ lieu_dit_uai              <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, N...
## $ boite_postale_uai         <chr> "519", "616", "157", "301", "302", "60309", "60309", NA, "303...
## $ code_postal_uai           <chr> "01505", "01206", "01306", "01011", "01011", "01011", "01011"...
## $ localite_acheminement_uai <chr> "AMBERIEU EN BUGEY CEDEX", "BELLEGARDE SUR VALSERINE C", "BEL...
## $ coordonnee_x              <chr> "881947,6", "917508", "909024,8", "871594,3", "871668,3", "87...
## $ coordonnee_y              <chr> "6544401,6", "6561249,4", "6521619,9", "6569319,7", "6568818,...
## $ appariement               <chr> "MANUEL", "MANUEL", "DIFF_NOM", "MANUEL", "MANUEL", "MANUEL",...
## $ localisation              <chr> "BATIMENT", "BATIMENT", "PLAQUE_ADRESSE", "BATIMENT", "BATIME...
## $ nature_uai                <chr> "320", "306", "300", "302", "300", "300", "320", "320", "320"...
## $ nature_uai_libe           <chr> "Lycée professionnel", "Lycée polyvalent", "Lycée d'enseignem...
## $ etat_etablissement        <chr> "1", "1", "1", "1", "1", "1", "1", "1", "1", "1", "1", "1", "...
}
}
\keyword{datasets}
