% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bourse}
\alias{bourse}
\title{Table bourse}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 6 rows and 2 columns.
}
\usage{
bourse
}
\description{
\preformatted{## Rows: 6
## Columns: 2
## $ code_bourse <chr> "01", "02", "03", "04", "05", "06"
## $ lib_bourse  <chr> "Française", "CROUS", "DEA", "DESS", "Agrégation", "CROUS"
}
}
\keyword{datasets}
