% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{compatibilite_mention_diplome_l}
\alias{compatibilite_mention_diplome_l}
\title{Renvoie les mentions master compatibles}
\usage{
compatibilite_mention_diplome_l(code_mention_diplome_l)
}
\arguments{
\item{code_mention_diplome_l}{Un vecteur de codes mentions licence.}
}
\value{
Une liste contenant les mentions master compatibles.
}
\description{
Renvoie les mentions master compatibles.
}
