% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{compatibilite_mention_diplome_m}
\alias{compatibilite_mention_diplome_m}
\title{Renvoie les mentions licence compatibles}
\usage{
compatibilite_mention_diplome_m(code_mention_diplome_m)
}
\arguments{
\item{code_mention_diplome_m}{Un vecteur de code mention master.}
}
\value{
Une liste contenant les mentions licence compatibles.
}
\description{
Renvoie les mentions licence compatibles.
}
