% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{composante}
\alias{composante}
\title{Table composante}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 55 rows and 7 columns.
}
\usage{
composante
}
\description{
\preformatted{## Rows: 55
## Columns: 7
## $ code_composante         <chr> "AL1", "AL2", "AL3", "ALB", "APS", "AUC", "BIO", "CAS", "CHI", ...
## $ lib_composante          <chr> "ALBI", "ALBI", "ALBI", "ALBI", "F2SMH", "IUT Paul Sabatier - A...
## $ code_commune_composante <chr> "81004", "81004", "81004", "81004", "31555", "32013", "31555", ...
## $ code_type_composante    <chr> "DPT", "LOC", "LOC", "LOC", "UFR", "LOC", "DPT", "LOC", "DPT", ...
## $ code_uai_composante     <chr> NA, "0811255Z", "0811198M", NA, "0311621U", "0320691C", NA, "08...
## $ code_composante_parent  <chr> NA, NA, NA, NA, NA, "IUTA", NA, "IUTA", NA, NA, NA, NA, NA, NA,...
## $ ville                   <chr> NA, NA, NA, NA, "Toulouse", "Auch", NA, "Castres", NA, "Toulous...
}
}
\keyword{datasets}
