% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{composante_histo}
\alias{composante_histo}
\title{Table composante_histo}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 21 rows and 3 columns.
}
\usage{
composante_histo
}
\description{
\preformatted{## Rows: 21
## Columns: 3
## $ code_composante      <chr> "DEU", "IU1", "IU2", "IU3", "IU4", "IU5", "IU6", "IU7", "IU8", "IU...
## $ code_composante_succ <chr> "FSI", "FSI", "FSI", "FSI", "FSI", "FSI", "FSI", "FSI", "FSI", "FS...
## $ date_maj             <date> 2018-07-10, 2018-07-10, 2018-07-10, 2018-07-10, 2018-07-10, 2018-...
}
}
\keyword{datasets}
