% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{composante_type}
\alias{composante_type}
\title{Table composante_type}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 7 rows and 2 columns.
}
\usage{
composante_type
}
\description{
\preformatted{## Rows: 7
## Columns: 2
## $ code_type_composante <chr> "COM", "DPT", "INS", "IUP", "IUT", "LOC", "UFR"
## $ lib_type_composante  <chr> "Coummunauté d'universités/établissements", "Département", "Instit...
}
}
\keyword{datasets}
