% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{conv_annee_etape_type_diplome}
\alias{conv_annee_etape_type_diplome}
\title{Renvoie le type de diplome a partir de l'annee d'etape}
\usage{
conv_annee_etape_type_diplome(annee_etape)
}
\arguments{
\item{code_elp}{Un vecteur d'année d'étape.}
}
\value{
Un vecteur contenant les acronymes de type de diplôme.

Jeu de données source : \code{apogee::diplome_type}.\cr
}
\description{
Renvoie le type de diplôme à partir de l'année d'étape.
}
