% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{conv_elp_periode}
\alias{conv_elp_periode}
\title{Renvoie le code de periode d'enseignement a partir du code ELP}
\usage{
conv_elp_periode(code_elp)
}
\arguments{
\item{code_elp}{Un vecteur de code ELP.}
}
\value{
Un vecteur contenant les codes de période d'enseignement.

Jeu de données source : \code{apogee::elp}.\cr
}
\description{
Renvoie le code de période d'enseignement à partir du code ELP.
}
