% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{conv_etape_discipline_sise}
\alias{conv_etape_discipline_sise}
\title{Renvoie le code de discipline SISE a partir du code etape}
\usage{
conv_etape_discipline_sise(code_etape)
}
\arguments{
\item{code_etape}{Un vecteur de code étape.}
}
\value{
Un vecteur contenant les codes de discipline SISE.

Jeu de données source : \code{apogee::etape_sise_discipline}.\cr
Il est créé à partir d'Apogée.
}
\description{
Renvoie le code de discipline SISE à partir du code étape.
}
