% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{conv_etape_finalite_diplome}
\alias{conv_etape_finalite_diplome}
\title{Renvoie le code de finalite diplome a partir du code etape}
\usage{
conv_etape_finalite_diplome(code_etape)
}
\arguments{
\item{code_etape}{Un vecteur de code étape.}
}
\value{
Un vecteur contenant les codes de finalité diplôme.

Jeu de données source : \code{apogee::etape_finalite}.\cr
}
\description{
Renvoie le code de finalité diplôme à partir du code étape.
}
