% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{conv_etape_sise_diplome}
\alias{conv_etape_sise_diplome}
\title{Renvoie le code de diplome SISE a partir du code etape}
\usage{
conv_etape_sise_diplome(code_etape, annee)
}
\arguments{
\item{code_etape}{Un vecteur de code étape.}
}
\value{
Un vecteur contenant les codes de diplôme SISE.

Jeu de données source : \code{apogee::etape_sise}.\cr
Il est créé à partir d'Apogée.
}
\description{
Renvoie le code de diplôme SISE à partir du code étape.
}
