% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{conv_etape_voie}
\alias{conv_etape_voie}
\title{Renvoie la voie a partir du code étape}
\usage{
conv_etape_voie(code_etape)
}
\arguments{
\item{code_etape}{Un vecteur de codes étape.}
}
\value{
Un vecteur contenant les voies.

Jeu de données source : \code{parcoursup::groupe_voie}.\cr
}
\description{
Renvoie la voie a partir du code étape
}
