% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_apogee.R
\name{data_individus_formation_origine}
\alias{data_individus_formation_origine}
\title{data_individus_formation_origine}
\usage{
data_individus_formation_origine()
}
\description{
data_individus_formation_origine
}
\keyword{internal}
