% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_apogee.R
\name{data_inscrits_peda}
\alias{data_inscrits_peda}
\title{data_inscrits_peda}
\usage{
data_inscrits_peda(derniere_annee = TRUE)
}
\arguments{
\item{derniere_annee}{\dots}
}
\description{
data_inscrits_peda
}
\keyword{internal}
