% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_apogee.R
\name{data_resultats_diplome}
\alias{data_resultats_diplome}
\title{data_resultats_diplome}
\usage{
data_resultats_diplome(derniere_annee = TRUE)
}
\arguments{
\item{derniere_annee}{\dots}
}
\description{
data_resultats_diplome
}
\keyword{internal}
