% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{departement_academie}
\alias{departement_academie}
\title{Table departement_academie}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 110 rows and 2 columns.
}
\usage{
departement_academie
}
\description{
\preformatted{## Rows: 110
## Columns: 2
## $ code_departement <chr> "975", "983", "984", "987", "989", "971", "972", "973", "974", "976", ...
## $ code_academie    <chr> "44", "40", "41", "42", "41", "32", "31", "33", "28", "43", "01", "24"...
}
}
\keyword{datasets}
