% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{diplome}
\alias{diplome}
\title{Table diplome}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2430 rows and 5 columns.
}
\usage{
diplome
}
\description{
\preformatted{## Rows: 2,430
## Columns: 5
## $ code_diplome        <chr> "1A001", "1A002", "1A008", "1A027", "1A071", "1A073", "1A096", "1AC...
## $ lib_diplome         <chr> "DEA MATH APPLIQUEES", "DEA MATH PURES", "DEA 2IL", "DEA MECA DES F...
## $ code_nature_diplome <chr> "1", "1", "1", "1", "1", "1", "1", "1", "1", "1", "1", "1", "1", "1...
## $ code_type_diplome   <chr> "DEA", "DEA", "DEA", "DEA", "DEA", "DEA", "DEA", "LMD/M2", "DEA", "...
## $ code_cycle_univ     <chr> "3", "3", "3", "3", "3", "3", "3", "3", "3", "3", "3", "3", "3", "3...
}
}
\keyword{datasets}
