% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{diplome_origine_type}
\alias{diplome_origine_type}
\title{Table diplome_origine_type}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 34 rows and 2 columns.
}
\usage{
diplome_origine_type
}
\description{
\preformatted{## Rows: 34
## Columns: 2
## $ code_type_diplome_origine <chr> "1", "2", "3", "A", "B", "C", "D", "E", "F", "G", "H", "I", "...
## $ lib_type_diplome_origine  <chr> "Diplôme d'établissement étranger secondaire", "Diplôme d'éta...
}
}
\keyword{datasets}
