% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{diplome_type}
\alias{diplome_type}
\title{Table diplome_type}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 67 rows and 5 columns.
}
\usage{
diplome_type
}
\description{
\preformatted{## Rows: 67
## Columns: 5
## $ code_type_diplome        <chr> "AEA dentai", "AFS", "AFSA", "Att. univ1", "Att. univ3", "Cap....
## $ lib_type_diplome         <chr> "Attestation d'études universitaires", "Attestation de formati...
## $ acronyme_type_diplome    <chr> "AEU dentaire", "AFS", "AFSA", "Att. univ1", "Att. univ3", "Ca...
## $ annee_etape              <int> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, 0, NA, NA,...
## $ code_type_diplome_parent <chr> "Santé", "Santé", "Santé", NA, "Santé", "Santé", "Santé", NA, ...
}
}
\keyword{datasets}
