% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{domaine_diplome}
\alias{domaine_diplome}
\title{Table domaine_diplome}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 21 rows and 3 columns.
}
\usage{
domaine_diplome
}
\description{
\preformatted{## Rows: 21
## Columns: 3
## $ code_domaine_diplome     <chr> "DL01", "DL02", "DL03", "DL04", "DL05", "DL06", "DL07", "DL08"...
## $ lib_domaine_diplome      <chr> "Mathématiques, informatique et applications", "Management des...
## $ acronyme_domaine_diplome <chr> NA, NA, NA, NA, NA, NA, "STAPS", NA, "STS", "STAPS", "SHS", "D...
}
}
\keyword{datasets}
