% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{elp}
\alias{elp}
\title{Table elp}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 65442 rows and 7 columns.
}
\usage{
elp
}
\description{
\preformatted{## Rows: 65,442
## Columns: 7
## $ code_elp         <chr> "093PA21", "1A0201M", "1A0202M", "1A0203M", "1A0204M", "1A0205M", "1A0...
## $ lib_elp_court    <chr> "Acquisition objectifs", "M11. SINGULARITE D'ENS.", "M12. SURFACES DE ...
## $ lib_elp          <chr> "Acquisition objectifs", "M11. SINGULARITE D'ENSEMBLES ANALYTIQUES", "...
## $ code_nature_elp  <chr> "MATI", "MODU", "MODU", "MODU", "MODU", "MODU", "MODU", "MODU", "MODU"...
## $ code_periode_elp <chr> "AN", "AN", "AN", "AN", "AN", "AN", "AN", "AN", "AN", "AN", "AN", "AN"...
## $ ects             <dbl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA...
## $ temoin_elp_ue    <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA...
}
}
\keyword{datasets}
