% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{elp_histo}
\alias{elp_histo}
\title{Table elp_histo}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 53 rows and 4 columns.
}
\usage{
elp_histo
}
\description{
\preformatted{## Rows: 53
## Columns: 4
## $ code_elp          <chr> "HLS304M", "HLSEM3I", "HLSEM4I", "HLSEM5I", "HLSEM6I", "HMSEM7I", "HM...
## $ code_elp_succ     <chr> "HLSAN3JM", "HLSAN3S", "HLSAN4S", "HLSAN5S", "HLSAN6S", "HMSAN7S", "H...
## $ date_maj          <date> 2017-10-18, 2017-10-18, 2017-10-18, 2017-10-18, 2017-10-18, 2017-10-...
## $ suppression_histo <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, N...
}
}
\keyword{datasets}
