% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{elp_nature}
\alias{elp_nature}
\title{Table elp_nature}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 44 rows and 2 columns.
}
\usage{
elp_nature
}
\description{
\preformatted{## Rows: 44
## Columns: 2
## $ code_nature_elp <chr> "1B08", "AN", "BLOC", "CERT", "CHOI", "CM", "ENTC", "ENTI", "MACO", "MA...
## $ lib_nature_elp  <chr> "REH - 1B08", "Année", "Bloc", "Certificat", "Elt à choi", "C.M.", "ENT...
}
}
\keyword{datasets}
