% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{elp_periode}
\alias{elp_periode}
\title{Table elp_periode}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 15 rows and 1 columns.
}
\usage{
elp_periode
}
\description{
\preformatted{## Rows: 15
## Columns: 1
## $ code_periode_elp <chr> "AN", "PN", "S", "S1", "S2", "S3", "S4", "S5", "S6", "S7", "S8", "S9",...
}
}
\keyword{datasets}
