% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{etablissement_type}
\alias{etablissement_type}
\title{Table etablissement_type}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 23 rows and 2 columns.
}
\usage{
etablissement_type
}
\description{
\preformatted{## Rows: 23
## Columns: 2
## $ code_type_etab <chr> "00", "01", "02", "03", "04", "05", "06", "07", "08", "09", "10", "11", ...
## $ lib_type_etab  <chr> "Université", "STS", "CPGE ou préparation intégrée", "Ecole de commerce,...
}
}
\keyword{datasets}
