% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{etape}
\alias{etape}
\title{Table etape}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2735 rows and 31 columns.
}
\usage{
etape
}
\description{
\preformatted{## Rows: 2,735
## Columns: 31
## $ code_etape                                <chr> "1ADFE1", "1AINT1", "1AMAA1", "1AMAF1", "1DIN...
## $ lib_etape                                 <chr> "M2R DYNAMIQUE DES FLUIDES, ENERGETIQUE ET TR...
## $ acronyme_etape                            <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, N...
## $ annee_etape                               <dbl> 5, 5, 5, 5, 2, 2, 5, 3, 3, 3, 3, 3, 3, 3, 4, ...
## $ annee_diplome                             <int> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, N...
## $ code_type_diplome                         <chr> "LMD/M2", "LMD/M2", "LMD/M2", "LMD/M2", "LMD/...
## $ option                                    <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, N...
## $ acronyme_option                           <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, N...
## $ particularite                             <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, N...
## $ acronyme_particularite                    <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, N...
## $ ville                                     <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, N...
## $ cohabilite                                <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, N...
## $ annees_activite                           <list> [<2004, 2005, 2006, 2007, 2008, 2009, 2010>,...
## $ actif                                     <lgl> FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FAL...
## $ n_inscrits                                <int> 34, 207, 86, 35, 296, 268, 31, 100, 197, NA, ...
## $ temoin_etape_apogee                       <lgl> TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRU...
## $ lib_etape_apogee                          <chr> "M2R DYNAMIQUE DES FLUIDES, ENERGETIQUE ET TR...
## $ lib_etape_option                          <chr> "M2R DYNAMIQUE DES FLUIDES, ENERGETIQUE ET TR...
## $ lib_etape_particularite                   <chr> "M2R DYNAMIQUE DES FLUIDES, ENERGETIQUE ET TR...
## $ lib_etape_ville                           <chr> "M2R DYNAMIQUE DES FLUIDES, ENERGETIQUE ET TR...
## $ lib_etape_option_particularite            <chr> "M2R DYNAMIQUE DES FLUIDES, ENERGETIQUE ET TR...
## $ lib_etape_ville_option                    <chr> "M2R DYNAMIQUE DES FLUIDES, ENERGETIQUE ET TR...
## $ lib_etape_ville_particularite             <chr> "M2R DYNAMIQUE DES FLUIDES, ENERGETIQUE ET TR...
## $ lib_etape_ville_option_particularite      <chr> "M2R DYNAMIQUE DES FLUIDES, ENERGETIQUE ET TR...
## $ acronyme_etape_option                     <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, N...
## $ acronyme_etape_option_particularite       <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, N...
## $ acronyme_etape_particularite              <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, N...
## $ acronyme_etape_ville                      <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, N...
## $ acronyme_etape_ville_option               <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, N...
## $ acronyme_etape_ville_option_particularite <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, N...
## $ acronyme_etape_ville_particularite        <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, N...
}
}
\keyword{datasets}
