% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{etape_annees_activite}
\alias{etape_annees_activite}
\title{Renvoie les annees d'activite d'une formation a partir du code etape}
\usage{
etape_annees_activite(code_etape, historique = FALSE)
}
\arguments{
\item{code_etape}{Un vecteur de code étape.}

\item{historique}{Témoin de prise en compte de l'historique des code_etape.}
}
\value{
Une liste contenant les années d'activité.

Jeu de données source : \code{apogee::etape}.\cr
}
\description{
Renvoie les années d'activité d'une formation à partir du code étape.
}
