% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{etape_composante}
\alias{etape_composante}
\title{Table etape_composante}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3836 rows and 4 columns.
}
\usage{
etape_composante
}
\description{
\preformatted{## Rows: 3,836
## Columns: 4
## $ code_etape      <chr> "1A0011", "1A0021", "1A0081", "1A0271", "1A0711", "1A0731", "1A0961", "...
## $ code_composante <chr> "MIG", "MIG", "MIG", "MIG", "MIG", "MIG", "MIG", "MIG", "MIG", "MIG", "...
## $ premiere_annee  <int> 1985, 1980, 1995, 1989, 1991, 1991, 1998, 2005, 2000, 2004, 1999, 1999,...
## $ derniere_annee  <int> 2003, 2003, 2003, 1998, 2003, 1998, 1999, 2006, 2003, 2010, 2003, 2003,...
}
}
\keyword{datasets}
