% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apogee.R
\name{etape_derniere_annee}
\alias{etape_derniere_annee}
\title{Renvoie la derniere annee d'une etape a partir du code etape}
\usage{
etape_derniere_annee(code_etape)
}
\arguments{
\item{code_etape}{Un vecteur de code étape.}
}
\value{
Un vecteur contenant les années.

Jeu de données source : \code{apogee::etape}.\cr
}
\description{
Renvoie la dernière annee d'une étape à partir du code étape.
}
