% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{etape_domaine}
\alias{etape_domaine}
\title{Table etape_domaine}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3800 rows and 2 columns.
}
\usage{
etape_domaine
}
\description{
\preformatted{## Rows: 3,800
## Columns: 2
## $ code_etape           <chr> "1A0011", "1A0021", "1A0081", "1A0271", "1A0711", "1A0731", "1A096...
## $ code_domaine_diplome <chr> "DM04", "DM04", "DM04", "DM05", "DM04", "DM04", "DM05", "DM04", "D...
}
}
\keyword{datasets}
