% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{etape_finalite}
\alias{etape_finalite}
\title{Table etape_finalite}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 806 rows and 2 columns.
}
\usage{
etape_finalite
}
\description{
\preformatted{## Rows: 806
## Columns: 2
## $ code_etape            <chr> "1ACST1", "1ADFE1", "1AIIH1", "1AINT1", "1AMAA1", "1AMAF1", "1ARC...
## $ code_finalite_diplome <chr> "002", "002", "002", "002", "002", "002", "002", "002", "002", "0...
}
}
\keyword{datasets}
