% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{etape_histo}
\alias{etape_histo}
\title{Table etape_histo}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 565 rows and 8 columns.
}
\usage{
etape_histo
}
\description{
\preformatted{## Rows: 565
## Columns: 8
## $ code_etape        <chr> "1ADFE1", "1AINT1", "1AMAA1", "1AMAF1", "1DINF1", "1DMME1", "1LINQ1",...
## $ code_elp          <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, N...
## $ code_etape_succ   <chr> "ERKDF1", "ERIIT1", "ERMFA1", "ERMFA1", "EDINF1", "EDMAT1", "ELINF1",...
## $ date_maj          <date> 2017-06-20, 2017-06-20, 2017-06-20, 2017-06-20, 2017-08-28, 2017-10-...
## $ suppression_histo <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, N...
## $ commentaire       <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, N...
## $ doublon           <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, N...
## $ doublon_elp       <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, N...
}
}
\keyword{datasets}
