% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{etape_mention}
\alias{etape_mention}
\title{Table etape_mention}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1372 rows and 3 columns.
}
\usage{
etape_mention
}
\description{
\preformatted{## Rows: 1,372
## Columns: 3
## $ code_etape           <chr> "1A0011", "1A0021", "1A0081", "1A0711", "1A0961", "1ACST1", "1ADF1...
## $ code_mention_diplome <chr> "MM015", "MM015", "MM012", "MM012", "MM016", "MM012", "MM016", "MM...
## $ date_maj             <date> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, N...
}
}
\keyword{datasets}
