% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apogee.R
\name{etape_premiere_annee}
\alias{etape_premiere_annee}
\title{Renvoie la premiere annee d'une etape a partir du code etape}
\usage{
etape_premiere_annee(code_etape, historique = FALSE)
}
\arguments{
\item{code_etape}{Un vecteur de code étape.}

\item{historique}{Témoin de prise en compte de l'historique des code_etape.}
}
\value{
Un vecteur contenant les années.

Jeu de données source : \code{apogee::etape}.\cr
}
\description{
Renvoie la première annee d'une étape à partir du code étape.
}
