% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{etape_secteur}
\alias{etape_secteur}
\title{Table etape_secteur}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 229 rows and 4 columns.
}
\usage{
etape_secteur
}
\description{
\preformatted{## Rows: 229
## Columns: 4
## $ code_etape      <chr> "3NAGP1", "ENSIA1", "2NIPA1", "INGPA1", "ENBQS1", "3NQSE1", "INGBD1", "...
## $ secteur         <chr> "Sciences et techniques", "Sciences et techniques", "Sciences et techni...
## $ categorie_scuio <chr> "Agriculture - Agronomie - Agroalimentaire", "Agriculture - Agronomie -...
## $ departement_dut <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, "GEII", NA, NA, NA,...
}
}
\keyword{datasets}
