% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{etape_sise}
\alias{etape_sise}
\title{Table etape_sise}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 13687 rows and 3 columns.
}
\usage{
etape_sise
}
\description{
\preformatted{## Rows: 13,687
## Columns: 3
## $ code_etape        <chr> "1ADFE1", "1ADFE1", "1ADFE1", "1ADFE1", "1AINT1", "1AINT1", "1AINT1",...
## $ annee             <int> 2007, 2008, 2009, 2010, 2007, 2008, 2009, 2010, 2007, 2008, 2009, 201...
## $ code_diplome_sise <chr> "2222989", "2222989", "2222989", "2222989", "2223851", "2223851", "22...
}
}
\keyword{datasets}
