% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{finalite_diplome}
\alias{finalite_diplome}
\title{Table finalite_diplome}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 4 rows and 2 columns.
}
\usage{
finalite_diplome
}
\description{
\preformatted{## Rows: 4
## Columns: 2
## $ code_finalite_diplome <chr> "001", "002", "004", "005"
## $ lib_finalite_diplome  <chr> "Professionnelle", "Recherche", "Indifférenciée", "Enseignement"
}
}
\keyword{datasets}
