% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flux.R
\name{flux_etape_succ}
\alias{flux_etape_succ}
\title{Renvoie le code etape successeur}
\usage{
flux_etape_succ(code_etape, successeur_final = FALSE,
  garder_na = FALSE)
}
\arguments{
\item{code_etape}{Un vecteur de code étape.}

\item{successeur_final}{\code{TRUE}, renvoit le successeur le plus récent dans l'historique; \code{FALSE}, renvoie le premier successeur.}

\item{garder_na}{\code{TRUE}, les codes sans successeur passent à \code{NA}; \code{FALSE}, les codes sans successeur sont gardés tels quels.}
}
\value{
Un vecteur de code étape successeur.

Jeu de données source : \code{apogee::etape_flux}.\cr
Il est créé à partir d'Apogée et de la table "etape_flux" de la base Access Tables_ref (projet Apogee).
}
\description{
Renvoie le code étape successeur.
}
