% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{formations_historique}
\alias{formations_historique}
\title{Historique des formations}
\usage{
formations_historique(annee_debut)
}
\arguments{
\item{annee_debut}{Année de début définie}
}
\description{
Historique des formations depuis l'année de début définie
}
