% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apogee.R
\name{formations_liste}
\alias{formations_liste}
\title{Liste des formations d'une annee universitaire}
\usage{
formations_liste(annee = NULL)
}
\arguments{
\item{annee}{Année (par défaut, celle en cours)}
}
\value{
Un tibble des formations
}
\description{
Liste des formations d'une année universitaire
}
