% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchie.R
\name{hier_bac_parent}
\alias{hier_bac_parent}
\title{Renvoie le code Bac parent}
\usage{
hier_bac_parent(code_bac, parent_final = FALSE, garder_na = FALSE)
}
\arguments{
\item{code_bac}{Un vecteur de code Bac.}

\item{parent_final}{\code{TRUE}, renvoit le parent le plus haut dans la hiérarchie; \code{FALSE}, renvoie le premier parent.}

\item{garder_na}{\code{TRUE}, les codes sans parent passent à \code{NA}; \code{FALSE}, les codes sans parent sont gardés tels quels.}
}
\value{
Un vecteur de code Bac parent.

Jeu de données source : \code{apogee::bac}.\cr
Il est créé à partir de la table "bac" de la base Access "Tables_ref.accdb" (projet Apogée).
}
\description{
Renvoie le code Bac parent.
}
