% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchie.R
\name{hier_composante_parent}
\alias{hier_composante_parent}
\title{Renvoie le code composante parent}
\usage{
hier_composante_parent(code_composante, parent_final = FALSE,
  garder_na = FALSE)
}
\arguments{
\item{code_composante}{Un vecteur de code composante.}

\item{parent_final}{\code{TRUE}, renvoit le parent le plus haut dans la hiérarchie; \code{FALSE}, renvoie le premier parent.}

\item{garder_na}{\code{TRUE}, les codes sans parent passent à \code{NA}; \code{FALSE}, les codes sans parent sont gardés tels quels.}
}
\value{
Un vecteur de code composante parent.

Jeu de données source : \code{apogee::composante}.\cr
Il est créé à partir de la table "composante" de la base Access "Tables_ref.accdb" (projet Apogée).
}
\description{
Renvoie le code composante parent.
}
