% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchie.R
\name{hier_departement_academie}
\alias{hier_departement_academie}
\title{Renvoie les codes academie a partir des codes de departement}
\usage{
hier_departement_academie(code_departement)
}
\arguments{
\item{code_departement}{Un vecteur de code de départements.}
}
\value{
Un vecteur de code académie.

Jeu de données source : \code{apogee::departement_academie}.\cr
Il est créé à partir de la table "departement_academie" de la base Access "Tables_ref.accdb".
}
\description{
Renvoie les codes académie à partir des codes de département.
}
\examples{
apogee::hier_departement_academie(c("031", "056"))
}
