% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchie.R
\name{hier_discipline_sise_parent}
\alias{hier_discipline_sise_parent}
\title{Renvoie le code discipline SISE parent}
\usage{
hier_discipline_sise_parent(code_discipline_sise, parent_final = FALSE,
  garder_na = FALSE)
}
\arguments{
\item{code_discipline_sise}{Un vecteur de code discipline SISE.}

\item{parent_final}{\code{TRUE}, renvoit le parent le plus haut dans la hiérarchie; \code{FALSE}, renvoie le premier parent.}

\item{garder_na}{\code{TRUE}, les codes sans parent passent à \code{NA}; \code{FALSE}, les codes sans parent sont gardés tels quels.}
}
\value{
Un vecteur de code discipline SISE parent.

Jeu de données source : \code{apogee::sise_discipline}.\cr
Il est créé à partir de la table "sise_discipline" de la base Access "Tables_ref.accdb" (projet Apogée).
}
\description{
Renvoie le code discipline SISE parent.
}
