% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchie.R
\name{hier_elp_nature}
\alias{hier_elp_nature}
\title{Renvoie la nature ELP a partir du code}
\usage{
hier_elp_nature(code_elp)
}
\arguments{
\item{code_elp}{Un vecteur de code ELP.}
}
\value{
Un vecteur contenant les natures ELP.
}
\description{
Renvoie la nature ELP a partir du code
}
