% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchie.R
\name{hier_etape_cycle}
\alias{hier_etape_cycle}
\title{Renvoie les codes de cycle a partir du code etape}
\usage{
hier_etape_cycle(code_etape)
}
\arguments{
\item{code_etape}{Un vecteur de code étape.}
}
\value{
Un vecteur de code cycle.

Jeu de données source : \code{apogee::etape_cycle}.\cr
}
\description{
Renvoie les codes de cycle à partir du code étape.
}
