% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchie.R
\name{hier_etape_domaine}
\alias{hier_etape_domaine}
\title{Renvoie le domaine de diplome a partir du code etape}
\usage{
hier_etape_domaine(code_etape)
}
\arguments{
\item{code_etape}{Un vecteur de code étape.}
}
\value{
Un vecteur contenant les domaines de diplôme.
}
\description{
Renvoie le domaine de diplôme à partir du code étape.
}
