% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchie.R
\name{hier_etape_filiere}
\alias{hier_etape_filiere}
\title{Renvoie la filiere a partir du code etape}
\usage{
hier_etape_filiere(code_etape)
}
\arguments{
\item{code_etape}{Un vecteur de code étape.}
}
\value{
Un vecteur contenant les filières.
}
\description{
Renvoie la filiere a partir du code etape
}
