% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchie.R
\name{hier_etape_mention}
\alias{hier_etape_mention}
\title{Renvoie la mention de diplome a partir du code etape}
\usage{
hier_etape_mention(code_etape)
}
\arguments{
\item{code_etape}{Un vecteur de code étape.}
}
\value{
Un vecteur contenant les mentions de diplôme.
}
\description{
Renvoie la mention de diplôme à partir du code étape.
}
