% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchie.R
\name{hier_etape_secteur}
\alias{hier_etape_secteur}
\title{Renvoie le secteur (secondaire/tertiaire) a partir du code etape}
\usage{
hier_etape_secteur(code_etape)
}
\arguments{
\item{code_etape}{Un vecteur de code étape.}
}
\value{
Un vecteur contenant les secteurs (secondaire/tertiaire).

Jeu de données source : \code{apogee::etape_secteur}.\cr
Il est créé à partir de la table "etape_secteur" de la base Access Tables_ref (projet Apogee).
}
\description{
Renvoie le secteur (secondaire/tertiaire) à partir du code étape.
}
