% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchie.R
\name{hier_etape_type_diplome}
\alias{hier_etape_type_diplome}
\title{Renvoie le type de diplome a partir du code etape}
\usage{
hier_etape_type_diplome(code_etape)
}
\arguments{
\item{code_etape}{Un vecteur de code étape.}
}
\value{
Un vecteur contenant les types de diplôme.
}
\description{
Renvoie le type de diplôme à partir du code étape.
}
