% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchie.R
\name{hier_pcs_parent}
\alias{hier_pcs_parent}
\title{Renvoie le code PCS parent}
\usage{
hier_pcs_parent(code_pcs, parent_final = FALSE, garder_na = FALSE)
}
\arguments{
\item{code_pcs}{Un vecteur de code PCS.}

\item{parent_final}{\code{TRUE}, renvoit le parent le plus haut dans la hiérarchie; \code{FALSE}, renvoie le premier parent.}

\item{garder_na}{\code{TRUE}, les codes sans parent passent à \code{NA}; \code{FALSE}, les codes sans parent sont gardés tels quels.}
}
\value{
Un vecteur de code PCS parent.

Jeu de données source : \code{apogee::pcs}.\cr
Il est créé à partir de la table "pcs" de la base Access "Tables_ref.accdb" (projet Apogée).
}
\description{
Renvoie le code PCS parent.
}
