% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchie.R
\name{hier_resultat_parent}
\alias{hier_resultat_parent}
\title{Renvoie le resultat parent}
\usage{
hier_resultat_parent(code_resultat)
}
\arguments{
\item{code_mention_diplome}{Un vecteur de résultats.}
}
\value{
Un vecteur contenant les résultats parent.
}
\description{
Renvoie le résultat parent.
}
