% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchie.R
\name{hier_type_diplome_parent}
\alias{hier_type_diplome_parent}
\title{Renvoie le type de diplome parent}
\usage{
hier_type_diplome_parent(code_type_diplome, parent_final = FALSE,
  garder_na = FALSE)
}
\arguments{
\item{code_mention_diplome}{Un vecteur de codes de type de diplôme.}
}
\value{
Un vecteur contenant les types de diplôme parent.
}
\description{
Renvoie le type de diplôme parent.
}
