% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/historique.R
\name{histo_composante_succ}
\alias{histo_composante_succ}
\title{Renvoie le code de composante successeur}
\usage{
histo_composante_succ(
  code_composante,
  successeur_final = TRUE,
  garder_na = FALSE
)
}
\arguments{
\item{code_composante}{Un vecteur de code de composante.}

\item{successeur_final}{\code{TRUE}, renvoit le successeur le plus récent dans l'historique; \code{FALSE}, renvoie le premier successeur.}

\item{garder_na}{\code{TRUE}, les codes sans successeur passent à \code{NA}; \code{FALSE}, les codes sans successeur sont gardés tels quels.}
}
\value{
Un vecteur de code de composante successeur.

Jeu de données source : \code{apogee::composante_histo}.\cr
Il est créé à partir d'Apogée et de la table "composante_histo" de la base Access Tables_ref (projet Apogee).
}
\description{
Renvoie le code de composante successeur.
}
