% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/historique.R
\name{histo_elp_succ}
\alias{histo_elp_succ}
\title{Renvoie le code ELP successeur}
\usage{
histo_elp_succ(code_elp, successeur_final = TRUE, garder_na = FALSE)
}
\arguments{
\item{code_elp}{Un vecteur de code ELP.}

\item{successeur_final}{\code{TRUE}, renvoit le successeur le plus récent dans l'historique; \code{FALSE}, renvoie le premier successeur.}

\item{garder_na}{\code{TRUE}, les codes sans successeur passent à \code{NA}; \code{FALSE}, les codes sans successeur sont gardés tels quels.}
}
\value{
Un vecteur de code ELP successeur.

Jeu de données source : \code{apogee::elp_histo}.\cr
Il est créé à partir d'Apogée et de la table "elp_histo" de la base Access Tables_ref (projet Apogee).
}
\description{
Renvoie le code ELP successeur.
}
