% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/historique.R
\name{histo_etape_pred}
\alias{histo_etape_pred}
\title{Renvoie les codes etape predecesseurs}
\usage{
histo_etape_pred(code_etape, annee = NULL, predecesseur_final = FALSE,
  garder_na = FALSE)
}
\arguments{
\item{code_etape}{Un vecteur de code étape.}

\item{predecesseur_final}{\code{TRUE}, renvoit le prédécesseur le plus ancien dans l'historique; \code{FALSE}, renvoie le premier prédécesseur.}

\item{garder_na}{\code{TRUE}, les codes sans prédécesseur passent à \code{NA}; \code{FALSE}, les codes sans successeur sont gardés tels quels.}
}
\value{
Une liste de vecteurs de code étape prédécesseur.

Jeu de données source : \code{apogee::etape_histo}.\cr
Il est créé à partir d'Apogée et de la table "etape_historique" de la base Access Tables_ref (projet Apogee).
}
\description{
Renvoie les codes étape prédécesseurs.
}
