% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/historique.R
\name{histo_mention_diplome_succ}
\alias{histo_mention_diplome_succ}
\title{Renvoie le code de mention de diplome successeur}
\usage{
histo_mention_diplome_succ(
  code_mention_diplome,
  successeur_final = TRUE,
  garder_na = FALSE
)
}
\arguments{
\item{successeur_final}{\code{TRUE}, renvoit le successeur le plus récent dans l'historique; \code{FALSE}, renvoie le premier successeur.}

\item{garder_na}{\code{TRUE}, les codes sans successeur passent à \code{NA}; \code{FALSE}, les codes sans successeur sont gardés tels quels.}

\item{code_etape}{Un vecteur de code de mention de diplôme.}
}
\value{
Un vecteur de code de mention de diplôme successeur.

Jeu de données source : \code{apogee::diplome_mention_histo}.\cr
Il est créé à partir d'Apogée et de la table "diplome_mention_histo" de la base Access Tables_ref (projet Apogee).
}
\description{
Renvoie le code de mention de diplôme successeur.
}
