% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{individus}
\alias{individus}
\title{Table individus}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 172625 rows and 17 columns.
}
\usage{
individus
}
\description{
\preformatted{## Rows: 172,625
## Columns: 17
## $ code_etudiant              <chr> "20000286", "20000313", "20000351", "20000393", "20000398", ...
## $ ine                        <chr> "1692154719H", "1693005364X", "1693019580Z", "1693008810T", ...
## $ nom                        <chr> "ARNOULD", "ESQUERRE", "DONADEL", "CABANAC", "VERGNES", "AZU...
## $ nom_usage                  <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, "LABAT ALLÉE", NA, "JAUZ...
## $ prenom                     <chr> "GUILLAUME", "ARIANE", "LORENE", "GUILLAUME", "VALERIE", "JA...
## $ sexe                       <chr> "M", "F", "F", "M", "F", "M", "F", "M", "F", "F", "M", "F", ...
## $ code_nationalite           <chr> "100", "100", "100", "100", "100", "100", "100", "100", "100...
## $ date_naissance             <date> 1980-11-28, 1982-11-08, 1982-10-19, 1982-03-08, 1982-03-04,...
## $ lib_ville_naissance        <chr> "PERPIGNAN", "TOULOUSE", "KOUROU", "TOULOUSE", "PARIS", "AUC...
## $ annee_bac                  <int> 1999, 2000, 2000, 2000, 2000, 2000, 2000, 2000, 2000, 2000, ...
## $ code_departement_bac       <chr> "065", "031", "031", "031", "012", "032", "031", "065", "081...
## $ code_etab_bac              <chr> "0650041S", NA, NA, NA, NA, "0320067Z", "0310024H", "0650026...
## $ code_mention_bac           <chr> "P", "B", "AB", "AB", "P", "AB", "B", "B", "AB", NA, "AB", "...
## $ code_bac                   <chr> "TIEL", "S", "SPC", "SMA", "S", "S", "SPC", "S", "S", "SMA",...
## $ code_type_etab_bac         <chr> "LY", "LY", "LY", "LY", "LY", "LY", "LY", "LY", "LY", "LY", ...
## $ email_ups                  <chr> "guillaume.arnould@univ-tlse3.fr", "ariane.esquerre@univ-tls...
## $ code_departement_naissance <chr> "066", "031", "973", "031", "075", "032", "031", "065", "075...
}
}
\keyword{datasets}
