% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{individus_diplome_origine}
\alias{individus_diplome_origine}
\title{Table individus_diplome_origine}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 211542 rows and 5 columns.
}
\usage{
individus_diplome_origine
}
\description{
\preformatted{## Rows: 211,542
## Columns: 5
## $ code_etudiant                         <chr> "20000286", "20000313", "20000313", "20000351", "...
## $ annee_diplome_obtenu                  <int> 2002, 2007, NA, 2011, 1999, 2004, NA, 2006, 2004,...
## $ code_type_diplome_origine             <chr> "C", "V", "K", "DS", "A", "MA", "K", "MA", "E", "...
## $ code_etab_diplome_obtenu              <chr> "0311384L", "0311384L", NA, NA, NA, "0311382J", N...
## $ code_departement_pays_dernier_diplome <chr> "031", "031", NA, "087", "012", "031", NA, "034",...
}
}
\keyword{datasets}
