% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{individus_situation_annee_precedente}
\alias{individus_situation_annee_precedente}
\title{Table individus_situation_annee_precedente}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 483441 rows and 3 columns.
}
\usage{
individus_situation_annee_precedente
}
\description{
\preformatted{## Rows: 483,441
## Columns: 3
## $ code_etudiant                   <chr> "20000286", "20000286", "20000286", "20000286", "200003...
## $ annee                           <int> 2000, 2001, 2002, 2016, 2000, 2001, 2002, 2003, 2004, 2...
## $ code_situation_annee_precedente <chr> "H", "C", "C", "U", "A", "H", "H", "H", "H", "H", "H", ...
}
}
\keyword{datasets}
