% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{inscrits}
\alias{inscrits}
\title{Table inscrits}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 452727 rows and 33 columns.
}
\usage{
inscrits
}
\description{
\preformatted{## Rows: 452,727
## Columns: 33
## $ annee                    <int> 2008, 2008, 2008, 2008, 2008, 2008, 2008, 2008, 2008, 2008, 20...
## $ code_etape               <chr> "1ADFE1", "1ADFE1", "1ADFE1", "1ADFE1", "1ADFE1", "1ADFE1", "1...
## $ code_etudiant            <chr> "20201350", "20401574", "20606715", "20608177", "20609317", "2...
## $ inscription_premiere     <lgl> TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TR...
## $ code_composante          <list> ["MIG", "MIG", "MIG", "MIG", "MIG", "MIG", "MIG", "MIG", "MIG...
## $ code_version_etape       <list> ["71", "71", "71", "71", "71", "71", "71", "71", "71", "71", ...
## $ date_inscription         <date> 2008-09-22, 2008-09-29, 2008-09-29, 2008-09-29, 2008-09-30, 2...
## $ date_annulation          <date> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, N...
## $ inscription_en_cours     <lgl> TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TR...
## $ inscription_resiliee     <lgl> FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE,...
## $ date_resiliation         <date> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, N...
## $ code_profil_etudiant     <chr> "NO", "NO", "NO", "NO", "NO", "NO", "NO", "NO", "NO", "NO", "N...
## $ code_regime_inscription  <chr> "1", "1", "1", "1", "1", "1", "1", "1", "1", "1", "1", "1", "1...
## $ vae                      <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA...
## $ code_statut_etudiant     <chr> "01", "01", "01", "01", "01", "01", "01", "01", "01", "01", "0...
## $ code_bourse              <chr> NA, NA, "02", NA, NA, NA, NA, "02", NA, "02", NA, NA, NA, NA, ...
## $ code_situation_sociale   <chr> "NO", "NO", "BO", "NO", "NO", "NO", "NO", "BO", "NO", "BO", "N...
## $ code_aide_financiere     <chr> "I", "I", "Q", "I", "K", "I", "N", "Q", "F", "Q", NA, "I", "I"...
## $ etape_diplomante         <chr> "O", "O", "O", "O", "O", "O", "O", "O", "O", "O", "O", "O", "O...
## $ sportif_haut_niveau      <lgl> FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE,...
## $ lib_type_handicap        <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA...
## $ code_echange             <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA...
## $ code_type_etab_parallele <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA...
## $ code_pcs_parent          <chr> "31", "76", "52", "31", "38", "37", "42", "47", "82", "76", "3...
## $ code_pcs_autre_parent    <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA...
## $ email                    <chr> NA, NA, NA, NA, "maroua_1985@yahoo.fr", NA, NA, NA, "ouattarab...
## $ tel_portable             <chr> "0615380820", "06 26 38 80 89", "0633481406", NA, "06 21 49 87...
## $ tel_adresse_annuelle     <chr> "0615380820", "06 26 38 80 89", "0633481406", "05 59 93 23 52"...
## $ tel_fixe                 <chr> "0561750113", "05 62 24 37 31", "05 63 33 98 74", "06 86 71 12...
## $ code_echelon_bourse      <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA...
## $ boursier                 <chr> "N", "N", "O", "N", "N", "N", "N", "O", "N", "O", "N", "N", "N...
## $ elp_parcours             <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA...
## $ ville                    <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA...
}
}
\keyword{datasets}
