% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{inscrits_annules}
\alias{inscrits_annules}
\title{Table inscrits_annules}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 30945 rows and 31 columns.
}
\usage{
inscrits_annules
}
\description{
\preformatted{## Rows: 30,945
## Columns: 31
## $ annee                    <int> 2008, 2008, 2008, 2008, 2008, 2008, 2008, 2008, 2008, 2008, 20...
## $ code_etudiant            <chr> "20000627", "20001139", "20002689", "20003473", "20003516", "2...
## $ date_inscription         <date> 2008-12-18, 2008-10-01, 2008-09-30, 2008-09-18, 2008-07-11, 2...
## $ date_annulation          <date> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, 2009-02-13, NA, N...
## $ inscription_en_cours     <lgl> FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE,...
## $ inscription_resiliee     <lgl> FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE,...
## $ date_resiliation         <date> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, N...
## $ code_etape               <chr> "3MBST2", "PMSAP2", "3MBST2", "1LMAP1", "CO0901", "3LSTP1", "C...
## $ inscription_premiere     <lgl> FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, TRUE, TRUE, FALSE, F...
## $ code_profil_etudiant     <chr> "NO", "DI", "DI", "NO", "PE", "NO", "PE", "NO", "IC", "NO", "F...
## $ code_regime_inscription  <chr> "1", "1", "1", "1", "1", "1", "1", "1", "1", "1", "2", "2", "2...
## $ vae                      <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA...
## $ code_statut_etudiant     <chr> "01", "04", "04", "01", "01", "01", "01", "01", "01", "01", "0...
## $ code_bourse              <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA...
## $ code_situation_sociale   <chr> "NO", "NO", "NO", "NO", "NO", "NO", "NO", "NO", "NO", "NO", "N...
## $ code_aide_financiere     <chr> "K", "K", "K", "I", "A", "I", "A", "F", "K", "A", "K", "K", "X...
## $ etape_diplomante         <chr> "O", "O", "O", "O", "O", "O", "O", "O", "O", "O", "O", "O", "O...
## $ sportif_haut_niveau      <lgl> FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE,...
## $ lib_type_handicap        <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA...
## $ code_echange             <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA...
## $ code_type_etab_parallele <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, "00", NA, ...
## $ code_pcs_parent          <chr> "73", "73", "52", "42", "73", "47", "45", "38", "53", "76", "4...
## $ code_pcs_autre_parent    <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA...
## $ email                    <chr> NA, NA, "benoitchaput@aol.com", NA, "frederic.roig@gmail.com",...
## $ tel_portable             <chr> "0672736837", "06.76.94.48.43", "06 72 39 17 70", "0625246118"...
## $ tel_adresse_annuelle     <chr> "0672736837", "06.76.94.48.43", "06 72 39 17 70", "0625246118"...
## $ tel_fixe                 <chr> "0672736837", "06.76.94.48.43", "06 72 39 17 70", "0546663737"...
## $ code_echelon_bourse      <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA...
## $ boursier                 <chr> "N", "N", "N", "N", "N", "N", "N", "N", "N", "N", "N", "N", "N...
## $ code_composante          <list> ["PUR", <"PUR", "RAN">, <"PUR", "RAN">, "MIG", "PCA", "SVT", ...
## $ code_version_etape       <list> ["73", <"73", "72">, <"73", "72">, "71", "971", "71", "971", ...
}
}
\keyword{datasets}
