% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{inscrits_cpge}
\alias{inscrits_cpge}
\title{Table inscrits_cpge}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 7171 rows and 33 columns.
}
\usage{
inscrits_cpge
}
\description{
\preformatted{## Rows: 7,171
## Columns: 33
## $ annee                    <int> 2008, 2008, 2015, 2015, 2015, 2015, 2015, 2015, 2015, 2015, 20...
## $ code_etape               <chr> "1DMME1", "2DPHY1", "ECPGE1", "ECPGE1", "ECPGE1", "ECPGE1", "E...
## $ code_etudiant            <chr> "20708304", "20708257", "21505544", "21509124", "21509125", "2...
## $ inscription_premiere     <lgl> TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TR...
## $ code_composante          <list> ["MIG", "PCA", "SCU", "SCU", "SCU", "SCU", "SCU", "SCU", "SCU...
## $ code_version_etape       <list> ["71", "71", "151", "151", "151", "151", "151", "151", "151",...
## $ date_inscription         <date> 2008-10-03, 2008-10-03, 2015-07-21, 2015-09-15, 2015-09-15, 2...
## $ date_annulation          <date> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, N...
## $ inscription_en_cours     <lgl> TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TR...
## $ inscription_resiliee     <lgl> FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE,...
## $ date_resiliation         <date> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, N...
## $ code_profil_etudiant     <chr> "CP", "CP", "NO", "CP", "CP", "CP", "CP", "CP", "NO", "NO", "C...
## $ code_regime_inscription  <chr> "1", "1", "1", "1", "1", "1", "1", "1", "1", "1", "1", "1", "1...
## $ vae                      <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA...
## $ code_statut_etudiant     <chr> "01", "01", "01", "01", "01", "01", "01", "01", "01", "01", "0...
## $ code_bourse              <chr> "02", "01", NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA...
## $ code_situation_sociale   <chr> "BO", "BO", "NO", "NO", "BC", "NO", "BC", "NO", "BC", "NO", "N...
## $ code_aide_financiere     <chr> "Q", "V", "I", "I", "Q", "I", "Q", "I", "Q", "I", "I", "Q", "I...
## $ etape_diplomante         <chr> "O", "O", "O", "O", "O", "O", "O", "O", "O", "O", "O", "O", "O...
## $ sportif_haut_niveau      <lgl> FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE,...
## $ lib_type_handicap        <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA...
## $ code_echange             <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA...
## $ code_type_etab_parallele <chr> "02", "02", "02", "02", "02", "02", "02", "02", "02", "02", "0...
## $ code_pcs_parent          <chr> "54", "73", "45", "47", "52", "38", "21", "38", "31", "48", "3...
## $ code_pcs_autre_parent    <chr> NA, NA, "37", "43", "61", "38", "34", "38", "46", "37", "42", ...
## $ email                    <chr> NA, NA, "alice.chaloupy@gmail.com", "gabriel.arrondeau@outlook...
## $ tel_portable             <chr> "06 87 07 05 26", "06 70 04 91 97", "06 95 29 99 76", "06 72 3...
## $ tel_adresse_annuelle     <chr> "06 87 07 05 26", "06 70 04 91 97", "05 61 30 00 85", "06 72 3...
## $ tel_fixe                 <chr> "05 61 70 42 78", "230 06 96 00 58", "05 61 30 00 85", "05 34 ...
## $ code_echelon_bourse      <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA...
## $ boursier                 <chr> "O", "O", "N", "N", "O", "N", "O", "N", "O", "N", "N", "O", "N...
## $ elp_parcours             <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA...
## $ ville                    <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA...
}
}
\keyword{datasets}
