% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{inscrits_elp}
\alias{inscrits_elp}
\title{Table inscrits_elp}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 7347815 rows and 5 columns.
}
\usage{
inscrits_elp
}
\description{
\preformatted{## Rows: 7,347,815
## Columns: 5
## $ annee                <int> 2012, 2012, 2012, 2012, 2012, 2012, 2012, 2012, 2012, 2012, 2012, ...
## $ code_etape           <chr> "2LBPS1", "2LBPS1", "2LBPS1", "2LBPS1", "2LBPS1", "2LBPS1", "2LBPS...
## $ code_etudiant        <chr> "20504472", "20504472", "20504472", "20504472", "20504472", "20504...
## $ inscription_premiere <lgl> TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, ...
## $ code_elp             <chr> "2L5BP11", "2L5BP12", "2L5BP13", "2L5BP14", "2L5BP1M", "2L5BP21", ...
}
}
\keyword{datasets}
