% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{inscrits_peda}
\alias{inscrits_peda}
\title{Table inscrits_peda}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 253589 rows and 3 columns.
}
\usage{
inscrits_peda
}
\description{
\preformatted{## Rows: 253,589
## Columns: 3
## $ annee         <int> 2012, 2012, 2012, 2012, 2012, 2012, 2012, 2012, 2012, 2012, 2012, 2012, 2...
## $ code_etape    <chr> "2LBPS1", "2LBPS1", "2LBPS1", "2LBPS1", "2LBPS1", "2LBPS1", "2LBPS1", "2L...
## $ code_etudiant <chr> "20504472", "20700142", "20700830", "20704570", "20707674", "20800195", "...
}
}
\keyword{datasets}
