% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libelle.R
\name{lib_academie}
\alias{lib_academie}
\title{Renvoie le libelle a partir du code academie}
\usage{
lib_academie(code_academie)
}
\arguments{
\item{code_academie}{Un vecteur de code académie.}
}
\value{
Un vecteur contenant les libellés d'académie.

Jeu de données source : \code{apogee::academie}.\cr
Il est créé à partir de la table "academie" de la base Access "Tables_ref.accdb".
}
\description{
Renvoie le libellé à partir du code académie.
}
\examples{
apogee::lib_academie(c("01", "09"))
}
