% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libelle.R
\name{lib_bac}
\alias{lib_bac}
\title{Renvoie le libelle a partir du code Bac}
\usage{
lib_bac(code_bac)
}
\arguments{
\item{code_bac}{Un vecteur de code Bac.}
}
\value{
Un vecteur contenant les libellés Bac.

Jeu de données source : \code{apogee::bac}.\cr
Il est créé à partir de la table "bac" de la base Access "Tables_ref.accdb" (projet Apogée).
}
\description{
Renvoie le libellé à partir du code Bac.
}
