% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libelle.R
\name{lib_bourse}
\alias{lib_bourse}
\title{Renvoie le libelle a partir du code bourse}
\usage{
lib_bourse(code_bourse)
}
\arguments{
\item{code_type_diplome}{Un vecteur de code de bourse.}
}
\value{
Un vecteur contenant les libellés de bourse.

Jeu de données source : \code{apogee::bourse}.\cr
}
\description{
Renvoie le libellé à partir du code bourse.
}
