% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libelle.R
\name{lib_composante}
\alias{lib_composante}
\title{Renvoie le libelle a partir du code composante}
\usage{
lib_composante(code_composante)
}
\arguments{
\item{code_composante}{Un vecteur de code composante.}
}
\value{
Un vecteur contenant les libellés composante.

Jeu de données source : \code{apogee::composante}.\cr
Il est créé à partir de la table "composante" de la base Access "Tables_ref.accdb" (projet Apogée).
}
\description{
Renvoie le libellé à partir du code composante.
}
