% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libelle.R
\name{lib_cursus_etape}
\alias{lib_cursus_etape}
\title{Renvoie le libelle a partir du code cursus d'une etape}
\usage{
lib_cursus_etape(code_cursus_etape)
}
\arguments{
\item{code_cursus_etape}{Un vecteur de code cursus d'une étape.}
}
\value{
Un vecteur contenant les libellé de cursus.

Jeu de données source : \code{apogee::cursus_etape}.\cr
}
\description{
Renvoie le libellé à partir du code cursus d'une étape.
}
