% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libelle.R
\name{lib_diplome}
\alias{lib_diplome}
\title{Renvoie le libelle a partir du code diplome}
\usage{
lib_diplome(code_diplome)
}
\arguments{
\item{code_diplome}{Un vecteur de code diplôme.}
}
\value{
Un vecteur contenant les libellés de diplôme.

Jeu de données source : \code{apogee::diplome}.\cr
}
\description{
Renvoie le libellé à partir du diplôme.
}
