% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libelle.R
\name{lib_diplome_sise_1}
\alias{lib_diplome_sise_1}
\title{Renvoie le libelle 1 a partir du code diplome SISE}
\usage{
lib_diplome_sise_1(code_diplome_sise)
}
\arguments{
\item{code_diplome}{Un vecteur de code diplôme SISE.}
}
\value{
Un vecteur contenant les libellés de diplôme SISE.

Jeu de données source : \code{apogee::sise_diplome_lib}.\cr
}
\description{
Renvoie le libellé 1 à partir du diplôme SISE.
}
