% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libelle.R
\name{lib_discipline_sise}
\alias{lib_discipline_sise}
\title{Renvoie le libelle a partir du code discipline SISE}
\usage{
lib_discipline_sise(code_discipline_sise)
}
\arguments{
\item{code_discipline_sise}{Un vecteur de code discipline SISE.}
}
\value{
Un vecteur contenant les libellés de discipline SISE

Jeu de données source : \code{apogee::sise_discipline}.\cr
Il est créé à partir de la table "sise_discipline" de la base Access "Tables_ref.accdb" (projet Apogée).
}
\description{
Renvoie le libellé à partir du code discipline SISE.
}
