% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libelle.R
\name{lib_domaine_diplome}
\alias{lib_domaine_diplome}
\title{Renvoie le libelle a partir du code domaine de diplome}
\usage{
lib_domaine_diplome(code_domaine_diplome)
}
\arguments{
\item{code_domaine_diplome}{Un vecteur de code domaine de diplôme.}
}
\value{
Un vecteur contenant les libellés de domaine de diplôme.

Jeu de données source : \code{apogee::diplome_domaine}.\cr
}
\description{
Renvoie le libellé à partir du code domaine de diplôme.
}
