% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libelle.R
\name{lib_elp}
\alias{lib_elp}
\title{Renvoie le libelle a partir du code ELP}
\usage{
lib_elp(code_elp)
}
\arguments{
\item{code_elp}{Un vecteur de code de ELP.}
}
\value{
Un vecteur contenant les libellés de ELP.

Jeu de données source : \code{apogee::elp}.\cr
}
\description{
Renvoie le libellé à partir du code ELP.
}
