% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libelle.R
\name{lib_finalite_diplome}
\alias{lib_finalite_diplome}
\title{Renvoie le libelle a partir du code de finalite de diplome}
\usage{
lib_finalite_diplome(code_finalite_diplome)
}
\arguments{
\item{code_finalite_diplome}{Un vecteur de code de finalité de diplôme.}
}
\value{
Un vecteur contenant les libellés de type diplôme.

Jeu de données source : \code{apogee::finalite_diplome}.\cr
}
\description{
Renvoie le libellé à partir du code de finalité de diplôme.
}
