% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libelle.R
\name{lib_mention_bac}
\alias{lib_mention_bac}
\title{Renvoie le libelle a partir du code de mention au Bac}
\usage{
lib_mention_bac(code_mention_bac)
}
\arguments{
\item{code_mention_bac}{Un vecteur de code de mention au Bac.}
}
\value{
Un vecteur contenant les libellés de mention au Bac.

Jeu de données source : \code{apogee::bac_mention}.\cr
Il est créé à partir de la table "bac_mention" de la base Access "Tables_ref.accdb" (projet Apogée).
}
\description{
Renvoie le libellé à partir du code de mention au Bac.
}
