% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libelle.R
\name{lib_mention_diplome}
\alias{lib_mention_diplome}
\title{Renvoie le libelle a partir du code mention de diplome}
\usage{
lib_mention_diplome(code_mention_diplome)
}
\arguments{
\item{code_mention_diplome}{Un vecteur de code mention de diplôme.}
}
\value{
Un vecteur contenant les libellés de mention de diplôme.

Jeu de données source : \code{apogee::diplome_mention}.\cr
}
\description{
Renvoie le libellé à partir du code mention de diplôme.
}
