% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libelle.R
\name{lib_nature_elp}
\alias{lib_nature_elp}
\title{Renvoie le libelle a partir du code nature ELP}
\usage{
lib_nature_elp(code_nature_elp)
}
\arguments{
\item{code_nature_elp}{Un vecteur de code de nature ELP.}
}
\value{
Un vecteur contenant les libellés de nature ELP.

Jeu de données source : \code{apogee::elp_nature}.\cr
}
\description{
Renvoie le libellé à partir du code nature ELP.
}
