% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libelle.R
\name{lib_pcs}
\alias{lib_pcs}
\title{Renvoie le libelle a partir du code PCS}
\usage{
lib_pcs(code_pcs)
}
\arguments{
\item{code_pcs}{Un vecteur de code PCS.}
}
\value{
Un vecteur contenant les libellés PCS.

Jeu de données source : \code{apogee::pcs}.\cr
Il est créé à partir de la table "pcs" de la base Access "Tables_ref.accdb" (projet Apogée).
}
\description{
Renvoie le libellé à partir du code PCS.
}
