% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libelle.R
\name{lib_periode_elp}
\alias{lib_periode_elp}
\title{Renvoie le libelle a partir du code de periode d'enseignement}
\usage{
lib_periode_elp(code_periode_elp)
}
\arguments{
\item{code_periode_elp}{Un vecteur de code de période d'enseignement.}
}
\value{
Un vecteur contenant les libellés de période d'enseignement.

Jeu de données source : \code{apogee::elp_periode}.\cr
}
\description{
Renvoie le libellé à partir code de période d'enseignement.
}
