% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libelle.R
\name{lib_regime_inscription}
\alias{lib_regime_inscription}
\title{Renvoie le libelle a partir du code de regime d'inscription}
\usage{
lib_regime_inscription(code_regime_inscription)
}
\arguments{
\item{code_regime_inscription}{Un vecteur de code de régime d'inscription.}
}
\value{
Un vecteur contenant les libellés de régime d'inscription.

Jeu de données source : \code{apogee::regime_inscription}.\cr
Il est créé à partir de la table "regime_inscription" de la base Access "Tables_ref.accdb" (projet Apogée).
}
\description{
Renvoie le libellé à partir du code de régime d'inscription.
}
