% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libelle.R
\name{lib_resultat}
\alias{lib_resultat}
\title{Renvoie le libelle a partir du code resultat}
\usage{
lib_resultat(code_resultat)
}
\arguments{
\item{code_resultat}{Un vecteur de code résultat.}
}
\value{
Un vecteur contenant les libellés de résultat.

Jeu de données source : \code{apogee::resultat}.\cr
}
\description{
Renvoie le libellé à partir du résultat.
}
