% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libelle.R
\name{lib_sexe}
\alias{lib_sexe}
\title{Renvoie le libelle a partir du code sexe}
\usage{
lib_sexe(code_sexe)
}
\arguments{
\item{code_sexe}{Un vecteur de code sexe.}
}
\value{
Un vecteur contenant les libellés de sexe.

Jeu de données source : \code{apogee::sexe}.\cr
Il est créé à partir de la table "sexe" de la base Access "Tables_ref.accdb" (projet Apogée).
}
\description{
Renvoie le libellé à partir du code sexe.
}
