% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libelle.R
\name{lib_sise_type_diplome}
\alias{lib_sise_type_diplome}
\title{Renvoie le libelle a partir du code de type de diplome SISE}
\usage{
lib_sise_type_diplome(code_sise_type_diplome)
}
\arguments{
\item{code_bac}{Un vecteur de code de type de diplôme SISE.}
}
\value{
Un vecteur contenant les libellés de type de diplôme SISE.

Jeu de données source : \code{apogee::sise_diplome_type}.\cr
Il est créé à partir de la table "sise_diplome_type" de la base Access "Tables_ref.accdb".
}
\description{
Renvoie le libellé à partir du code de type de diplôme SISE.
}
