% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libelle.R
\name{lib_situation_annee_precedente}
\alias{lib_situation_annee_precedente}
\title{Renvoie le libelle a partir du code de la situation l'annee precedente}
\usage{
lib_situation_annee_precedente(code_situation_annee_precedente)
}
\arguments{
\item{code_situation_annee_precedente}{Un vecteur de code de la situation l'année précédente.}
}
\value{
Un vecteur contenant les libellé de la situation l'année précédente.

Jeu de données source : \code{apogee::situation_annee_precedente}.\cr
}
\description{
Renvoie le libellé à partir du code de la situation l'année précédente.
}
