% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libelle.R
\name{lib_situation_sociale}
\alias{lib_situation_sociale}
\title{Renvoie le libelle a partir du code situation sociale}
\usage{
lib_situation_sociale(code_situation_sociale)
}
\arguments{
\item{code_situation_sociale}{Un vecteur de code situation sociale.}
}
\value{
Un vecteur contenant les libellés de situation sociale.

Jeu de données source : \code{apogee::situation_sociale}.\cr
}
\description{
Renvoie le libellé à partir du code situation sociale.
}
