% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libelle.R
\name{lib_specialite_diplome}
\alias{lib_specialite_diplome}
\title{Renvoie le libelle a partir du code de specialite diplome}
\usage{
lib_specialite_diplome(code_specialite_diplome)
}
\arguments{
\item{code_etape}{Un vecteur de code de spécialité diplôme.}
}
\value{
Un vecteur contenant les libellés de spécialité diplôme.

Jeu de données source : \code{apogee::diplome_specialite}.\cr
}
\description{
Renvoie le libellé à partir du code de spécialité diplôme.
}
