% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libelle.R
\name{lib_statut_etudiant}
\alias{lib_statut_etudiant}
\title{Renvoie le libelle a partir du code statut etudiant}
\usage{
lib_statut_etudiant(code_statut_etudiant)
}
\arguments{
\item{code_statut_etudiant}{Un vecteur de code statut étudiant.}
}
\value{
Un vecteur contenant les libellés de statut étudiant.

Jeu de données source : \code{apogee::statut_etudiant}.\cr
}
\description{
Renvoie le libellé à partir du code statut étudiant.
}
