% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libelle.R
\name{lib_type_diplome}
\alias{lib_type_diplome}
\title{Renvoie le libelle a partir du code de type diplome}
\usage{
lib_type_diplome(code_type_diplome)
}
\arguments{
\item{code_type_diplome}{Un vecteur de code de type diplôme.}
}
\value{
Un vecteur contenant les libellés de type diplôme.

Jeu de données source : \code{apogee::diplome_type}.\cr
Il est créé à partir de la table "diplome_type" de la base Access "Tables_ref.accdb" (projet Apogée).
}
\description{
Renvoie le libellé à partir du code de type diplôme.
}
