% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libelle.R
\name{lib_type_diplome_anterieur}
\alias{lib_type_diplome_anterieur}
\title{Renvoie le libelle a partir du code de type diplome anterieur}
\usage{
lib_type_diplome_anterieur(code_type_diplome_anterieur)
}
\arguments{
\item{code_type_diplome_anterieur}{Un vecteur de code de type diplôme antérieur.}
}
\value{
Un vecteur contenant les libellés de type diplôme antérieur.

Jeu de données source : \code{apogee::diplome_anterieur_type}.\cr
Il est créé à partir de l'objet "Type dernier diplôme obtenu" d'Apogée (code et lib).
}
\description{
Renvoie le libellé à partir du code de type diplôme antérieur.
}
