% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libelle.R
\name{lib_type_diplome_externe}
\alias{lib_type_diplome_externe}
\title{Renvoie le libelle a partir du code de type diplome externe}
\usage{
lib_type_diplome_externe(code_type_diplome_externe)
}
\arguments{
\item{code_type_diplome_externe}{Un vecteur de code de type diplôme externe.}
}
\value{
Un vecteur contenant les libellés de type diplôme externe.

Jeu de données source : \code{apogee::diplome_externe_type}.\cr
Il est créé à partir de l'objet "Type de diplôme externe" d'Apogée (code et lib).
}
\description{
Renvoie le libellé à partir du code de type diplôme externe.
}
