% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libelle.R
\name{lib_type_diplome_origine}
\alias{lib_type_diplome_origine}
\title{Renvoie le libelle a partir du code de type diplome d'origine}
\usage{
lib_type_diplome_origine(code_type_diplome_origine)
}
\arguments{
\item{code_type_diplome_origine}{Un vecteur de code de type diplôme d'origine}
}
\value{
Un vecteur contenant les libellés de type diplôme d'origine

Jeu de données source : \code{apogee::diplome_origine_type}.\cr
}
\description{
Renvoie le libellé à partir du code de type diplôme d'origine
}
