% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libelle.R
\name{lib_type_etab}
\alias{lib_type_etab}
\title{Renvoie le libelle a partir du code type etablissement}
\usage{
lib_type_etab(code_type_etab)
}
\arguments{
\item{code_type_etab}{Un vecteur de code de type établissement.}
}
\value{
Un vecteur contenant les libellés de type établissement.

Jeu de données source : \code{apogee::etablissement_type}.\cr
}
\description{
Renvoie le libellé à partir du code type établissement.
}
