% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apogee.R
\name{liste_formations}
\alias{liste_formations}
\title{Liste des formations d'une annee universitaire}
\usage{
liste_formations(annee)
}
\arguments{
\item{annee}{Année}
}
\value{
Un tibble des fromations
}
\description{
Liste des formations d'une année universitaire
}
