% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mention_diplome}
\alias{mention_diplome}
\title{Table mention_diplome}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 183 rows and 7 columns.
}
\usage{
mention_diplome
}
\description{
\preformatted{## Rows: 183
## Columns: 7
## $ code_mention_diplome        <chr> "ML001", "ML002", "ML003", "ML004", "ML005", "ML006", "ML00...
## $ lib_mention_diplome         <chr> "Activités physiques adaptées", "Biochimie et biologie molé...
## $ acronyme_mention_diplome    <chr> NA, NA, NA, NA, NA, "CHIMIE-L", NA, NA, NA, NA, NA, "INFOCO...
## $ code_mention_diplome_parent <chr> NA, NA, NA, NA, NA, "LSF", NA, NA, NA, NA, NA, NA, "LSN", N...
## $ mention_premiere_annee      <int> 2008, 2008, 2008, 2008, 2008, 2008, NA, 2008, 2008, 2008, 2...
## $ mention_derniere_annee      <int> 2010, 2010, 2010, 2010, 2010, 2019, NA, 2010, 2015, 2010, 2...
## $ actif                       <lgl> FALSE, FALSE, FALSE, FALSE, FALSE, TRUE, FALSE, FALSE, FALS...
}
}
\keyword{datasets}
