% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mention_diplome_histo}
\alias{mention_diplome_histo}
\title{Table mention_diplome_histo}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 4 rows and 3 columns.
}
\usage{
mention_diplome_histo
}
\description{
\preformatted{## Rows: 4
## Columns: 3
## $ code_mention_diplome      <chr> "ML030", "MM009", "MM027", "MM047"
## $ code_mention_diplome_succ <chr> "ML044", "MM027", "MM071", "MM075"
## $ date_maj                  <date> NA, NA, NA, 2019-02-15
}
}
\keyword{datasets}
