% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mention_diplome_lm}
\alias{mention_diplome_lm}
\title{Table mention_diplome_lm}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 60 rows and 2 columns.
}
\usage{
mention_diplome_lm
}
\description{
\preformatted{## Rows: 60
## Columns: 2
## $ code_mention_diplome_l <chr> "ML032", "ML012", "ML032", "ML032", "ML049", "ML049", "ML016", "...
## $ code_mention_diplome_m <chr> "MM036", "MM011", "MM074", "MM062", "MM062", "MM015", "MM012", "...
}
}
\keyword{datasets}
