% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_apogee.R
\name{nest_inscrits}
\alias{nest_inscrits}
\title{nest_inscrits}
\usage{
nest_inscrits(table, champ_nest, cle = c("annee", "code_etape",
  "code_etudiant", "inscription_premiere"))
}
\arguments{
\item{table}{\dots}

\item{champ_nest}{\dots}

\item{cle}{\dots}
}
\description{
nest_inscrits
}
\keyword{internal}
