% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pcs}
\alias{pcs}
\title{Table pcs}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 557 rows and 3 columns.
}
\usage{
pcs
}
\description{
\preformatted{## Rows: 557
## Columns: 3
## $ code_pcs        <chr> "10", "21", "22", "23", "31", "32", "36", "41", "46", "47", "48", "51",...
## $ lib_pcs         <chr> "Agriculteurs exploitants", "Artisans", "Commerçants et assimilés", "Ch...
## $ code_pcs_parent <chr> "1", "2", "2", "2", "3", "3", "3", "4", "4", "4", "4", "5", "5", "5", "...
}
}
\keyword{datasets}
