% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{profil_etudiant}
\alias{profil_etudiant}
\title{Table profil_etudiant}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 59 rows and 2 columns.
}
\usage{
profil_etudiant
}
\description{
\preformatted{## Rows: 59
## Columns: 2
## $ code_profil_etudiant <chr> "2D", "2E", "AE", "AL", "AS", "AU", "CA", "CC", "CD", "CE", "CH", ...
## $ lib_profil_etudiant  <chr> "UPS D2E", "Etud. entr", "Emp.Jeune", "auditeur", "au-disport", "a...
}
}
\keyword{datasets}
