% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{regime_inscription}
\alias{regime_inscription}
\title{Table regime_inscription}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 6 rows and 3 columns.
}
\usage{
regime_inscription
}
\description{
\preformatted{## Rows: 6
## Columns: 3
## $ code_regime_inscription  <chr> "1", "2", "3", "4", "5", "7"
## $ lib_regime_inscription   <chr> "Initiale", "Continue", "Reprise d'études", "Apprentissage", "...
## $ ordre_regime_inscription <int> 1, 4, 5, 2, 6, 3
}
}
\keyword{datasets}
