% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{resultat}
\alias{resultat}
\title{Table resultat}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 48 rows and 3 columns.
}
\usage{
resultat
}
\description{
\preformatted{## Rows: 48
## Columns: 3
## $ code_resultat        <chr> "ABI", "ABJ", "ACES", "ADAC", "ADC", "ADJ", "ADM", "ADMI", "ADMP",...
## $ lib_resultat         <chr> "Abs injustifiée", "Abs justifiée", "Ann. césure", "Admis avt choi...
## $ code_resultat_parent <chr> "DEF_DEM", "DEF_DEM", "CES", "ADM", "ADM", "ADM", NA, "ADM", "ADM"...
}
}
\keyword{datasets}
