% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{resultats_diplome}
\alias{resultats_diplome}
\title{Table resultats_diplome}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 165153 rows and 12 columns.
}
\usage{
resultats_diplome
}
\description{
\preformatted{## Rows: 165,153
## Columns: 12
## $ annee                <int> 2008, 2008, 2008, 2008, 2008, 2008, 2008, 2008, 2008, 2008, 2008, ...
## $ code_etape           <chr> "1ADFE1", "1ADFE1", "1ADFE1", "1ADFE1", "1ADFE1", "1ADFE1", "1ADFE...
## $ code_etudiant        <chr> "20201350", "20401574", "20606715", "20608177", "20608177", "20609...
## $ inscription_premiere <lgl> TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, ...
## $ lib_session          <chr> "Session 1", "Session 1", "Session 1", "Session 1", "Session 2", "...
## $ presence_examen      <chr> "O", "O", "O", "N", "O", "O", "O", "O", "O", "O", "O", "N", "N", "...
## $ note                 <dbl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA...
## $ code_resultat        <chr> "ADM", "ADM", "ADM", "DEF", "ADM", "AJ", "AJ", "AJ", "AJ", "AJ", "...
## $ code_mention         <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA...
## $ session              <chr> "Session 1", "Session 1", "Session 1", "Session 2", "Session 2", N...
## $ diplome_obtention    <chr> "O", "O", "O", "O", "O", "N", "N", "N", "N", "O", "O", "N", "N", "...
## $ ects                 <dbl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA...
}
}
\keyword{datasets}
