% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{resultats_elp}
\alias{resultats_elp}
\title{Table resultats_elp}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 6064634 rows and 10 columns.
}
\usage{
resultats_elp
}
\description{
\preformatted{## Rows: 6,064,634
## Columns: 10
## $ annee                <int> 2012, 2012, 2012, 2012, 2012, 2012, 2012, 2012, 2012, 2012, 2012, ...
## $ code_etape           <chr> "CUC2E1", "CUC2E1", "CUC2E1", "CUC2E1", "CUC2E1", "CUC2E1", "CUC2E...
## $ code_etudiant        <chr> "20000690", "20000690", "20000690", "20000690", "20000690", "20000...
## $ inscription_premiere <lgl> FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FAL...
## $ code_elp             <chr> "CUC2A1", "CUC2A11", "CUC2A12", "CUC2A13", "CUC2A14", "CUC2A15", "...
## $ lib_session          <chr> "Session 1", "Session 1", "Session 1", "Session 1", "Session 1", "...
## $ code_resultat        <chr> "ADM", "AJ", "AJ", "ADM", "ADM", "ADM", "ADM", "ADM", "ADM", "ADM"...
## $ note                 <dbl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA...
## $ ects                 <dbl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA...
## $ present_examen       <lgl> TRUE, FALSE, FALSE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE...
}
}
\keyword{datasets}
