% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{resultats_etape}
\alias{resultats_etape}
\title{Table resultats_etape}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 396593 rows and 10 columns.
}
\usage{
resultats_etape
}
\description{
\preformatted{## Rows: 396,593
## Columns: 10
## $ annee                <int> 2008, 2008, 2008, 2008, 2008, 2008, 2008, 2008, 2008, 2008, 2008, ...
## $ code_etape           <chr> "1ADFE1", "1ADFE1", "1ADFE1", "1ADFE1", "1ADFE1", "1ADFE1", "1ADFE...
## $ code_etudiant        <chr> "20201350", "20401574", "20606715", "20608177", "20608177", "20609...
## $ inscription_premiere <lgl> TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, ...
## $ lib_session          <chr> "Session 1", "Session 1", "Session 1", "Session 1", "Session 2", "...
## $ note                 <dbl> 15.185, 16.519, 13.627, NA, 12.709, 9.394, 9.194, 9.902, 9.902, 12...
## $ code_resultat        <chr> "ADM", "ADM", "ADM", "DEF", "ADM", "AJ", "AJ", "AJ", "AJ", "AJ", "...
## $ code_mention         <chr> "B", "TB", "AB", NA, "AB", NA, NA, NA, NA, NA, "AB", NA, NA, NA, "...
## $ ects                 <dbl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA...
## $ code_elp             <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA...
}
}
\keyword{datasets}
