% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sexe}
\alias{sexe}
\title{Table sexe}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2 rows and 2 columns.
}
\usage{
sexe
}
\description{
\preformatted{## Rows: 2
## Columns: 2
## $ code_sexe <chr> "M", "F"
## $ lib_sexe  <chr> "Homme", "Femme"
}
}
\keyword{datasets}
