% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{situation_annee_precedente}
\alias{situation_annee_precedente}
\title{Table situation_annee_precedente}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 19 rows and 2 columns.
}
\usage{
situation_annee_precedente
}
\description{
\preformatted{## Rows: 19
## Columns: 2
## $ code_situation_annee_precedente <chr> "A", "B", "C", "D", "E", "F", "G", "H", "J", "K", "L", ...
## $ lib_situation_annee_precedente  <chr> "Enseignement secondaire (y compris par correspondance)...
}
}
\keyword{datasets}
