% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{situation_sociale}
\alias{situation_sociale}
\title{Table situation_sociale}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 9 rows and 2 columns.
}
\usage{
situation_sociale
}
\description{
\preformatted{## Rows: 9
## Columns: 2
## $ code_situation_sociale <chr> "BC", "BO", "CH", "DD", "NO", "PN", "PO", "SN", "TH"
## $ lib_situation_sociale  <chr> "Bours CPGE", "Boursier", "Chomeur", "Demi Droit", "Normal", "Pu...
}
}
\keyword{datasets}
