% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{statut_etudiant}
\alias{statut_etudiant}
\title{Table statut_etudiant}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 7 rows and 2 columns.
}
\usage{
statut_etudiant
}
\description{
\preformatted{## Rows: 7
## Columns: 2
## $ code_statut_etudiant <chr> "01", "02", "03", "04", "05", "06", "07"
## $ lib_statut_etudiant  <chr> "Etudiant", "Audit. lib", "Form. cont", "internat", "FC avec SS", ...
}
}
\keyword{datasets}
