% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apogee.R
\name{temoin_annee1_diplome}
\alias{temoin_annee1_diplome}
\title{Temoin TRUE/FALSE si le code etape est la premiere annee d'un diplome}
\usage{
temoin_annee1_diplome(code_etape)
}
\arguments{
\item{code_etape}{Un vecteur de code étape.}
}
\value{
Un vecteur de booléens TRUE/FALSE.
}
\description{
Témoin TRUE/FALSE si le code étape est la première année d'un diplôme.
}
