% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apogee.R
\name{temoin_elp_ue}
\alias{temoin_elp_ue}
\title{Temoin TRUE/FALSE si l'ELP est une UE}
\usage{
temoin_elp_ue(code_elp)
}
\arguments{
\item{code_elp}{Un vecteur de code ELP.}
}
\value{
Un vecteur de booléens TRUE/FALSE.
}
\description{
Témoin TRUE/FALSE si l'ELP est une UE
}
