% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{temoin_mention_actif}
\alias{temoin_mention_actif}
\title{Temoin TRUE/FALSE si le code_mention_diplome est actif}
\usage{
temoin_mention_actif(code_mention_diplome, annee = NULL)
}
\arguments{
\item{code_mention_diplome}{Un vecteur de code_mention_diplome.}

\item{annee}{Année(s) d'activité à vérifier, par défaut l'année en cours.}
}
\value{
Un vecteur de booléens TRUE/FALSE.
}
\description{
Témoin TRUE/FALSE si le code_mention_diplome est actif.
}
