% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libelle.R
\name{acronyme_etape}
\alias{acronyme_etape}
\title{Renvoie l'acronyme a partir du code etape}
\usage{
acronyme_etape(code_etape)
}
\arguments{
\item{code_etape}{Un vecteur de code étape.}
}
\value{
Un vecteur contenant les sigle étape.

Jeu de données source : \code{apogee::etape}.\cr
Il est créé à partir de la table "etape" de la base Access "Tables_ref.accdb" (projet Apogée).
}
\description{
Renvoie l'acronyme à partir du code étape.
}
