% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apogee.R
\name{annee_en_cours}
\alias{annee_en_cours}
\title{Retourne l'année universitaire en cours}
\usage{
annee_en_cours(mois_resultats = 10)
}
\arguments{
\item{mois_resultats}{Mois des résultats, octobre par défaut}
}
\value{
Une valeur integer contenant l'année en cours
}
\description{
Retourne l'année universitaire en cours selon le mois défini pour les résultats
}
