% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{annee_u}
\alias{annee_u}
\title{Transformation de l'annee d'inscription en annee universitaire}
\usage{
annee_u(annee, fichier = FALSE)
}
\arguments{
\item{annee}{Année d'inscription.}

\item{fichier}{Si \code{TRUE}, utilisation du trait d'union et non du tiret insécable.}
}
\value{
Année universitaire.
}
\description{
Transformation de l'année d'inscription en année universitaire.
}
\examples{

apogee::annee_u(2016)
apogee::annee_u(2016, fichier = TRUE)

}
