% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{etape}
\alias{etape}
\alias{etape_composante}
\title{Tables liées aux etapes}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3643 rows and 40 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 3836 rows and 4 columns.
}
\usage{
etape

etape_composante
}
\description{
Tables liées aux etapes

\preformatted{## Observations: 3,643
## Variables: 40
## $ code_etape                                <chr> "1A0011", "1A0021", "1A0081", "1A0271", "1A07...
## $ lib_etape_court                           <chr> "DEA MATH APPLIQUEES", "DEA MATH PURES", "DEA...
## $ lib_etape_apogee                          <chr> "DEA MATHEMATIQUES APPLIQUEES", "DEA MATHEMAT...
## $ annee1_diplome                            <chr> "O", "O", "O", "O", "O", "O", "O", "N", "O", ...
## $ lib_etape                                 <chr> "DEA MATHEMATIQUES APPLIQUEES", "DEA MATHEMAT...
## $ acronyme_etape                            <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, N...
## $ actif                                     <lgl> FALSE, FALSE, FALSE, FALSE, FALSE, FALSE, FAL...
## $ annee_etape                               <dbl> 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, ...
## $ annee_diplome                             <int> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, N...
## $ filiere                                   <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, N...
## $ option                                    <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, N...
## $ acronyme_option                           <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, N...
## $ particularite                             <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, N...
## $ acronyme_particularite                    <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, N...
## $ ville                                     <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, N...
## $ cohabilite                                <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, N...
## $ commentaire                               <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, N...
## $ date_maj                                  <date> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, ...
## $ temoin_access                             <lgl> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, N...
## $ code_type_diplome                         <chr> "DEA", "DEA", "DEA", "DEA", "DEA", "DEA", "DE...
## $ lib_etape_option                          <chr> "DEA MATHEMATIQUES APPLIQUEES", "DEA MATHEMAT...
## $ lib_etape_particularite                   <chr> "DEA MATHEMATIQUES APPLIQUEES", "DEA MATHEMAT...
## $ lib_etape_ville                           <chr> "DEA MATHEMATIQUES APPLIQUEES", "DEA MATHEMAT...
## $ lib_etape_option_particularite            <chr> "DEA MATHEMATIQUES APPLIQUEES", "DEA MATHEMAT...
## $ lib_etape_ville_option                    <chr> "DEA MATHEMATIQUES APPLIQUEES", "DEA MATHEMAT...
## $ lib_etape_ville_particularite             <chr> "DEA MATHEMATIQUES APPLIQUEES", "DEA MATHEMAT...
## $ lib_etape_ville_option_particularite      <chr> "DEA MATHEMATIQUES APPLIQUEES", "DEA MATHEMAT...
## $ annee_diplome_apogee                      <int> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, N...
## $ acronyme_etape_option                     <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, N...
## $ acronyme_etape_option_particularite       <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, N...
## $ acronyme_etape_particularite              <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, N...
## $ acronyme_etape_ville                      <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, N...
## $ acronyme_etape_ville_option               <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, N...
## $ acronyme_etape_ville_option_particularite <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, N...
## $ acronyme_etape_ville_particularite        <chr> NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, N...
## $ temoin_etape_apogee                       <lgl> TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRUE, TRU...
## $ n_inscrits                                <int> NA, NA, NA, NA, NA, NA, NA, NA, NA, 34, NA, N...
## $ annee_premiere_etape                      <int> NA, NA, NA, NA, NA, NA, NA, NA, NA, 2008, NA,...
## $ annee_derniere_etape                      <int> NA, NA, NA, NA, NA, NA, NA, NA, NA, 2010, NA,...
## $ annee_etape_apogee                        <dbl> 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, 5, ...
}
}
\keyword{datasets}
