% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchie.R
\name{hier_mention_parent}
\alias{hier_mention_parent}
\title{Renvoie la mention de diplome parent}
\usage{
hier_mention_parent(code_mention_diplome, parent_final = FALSE,
  garder_na = FALSE)
}
\arguments{
\item{code_mention_diplome}{Un vecteur de code de mention de diplôme.}

\item{parent_final}{\code{TRUE}, renvoit le parent le plus haut dans la hiérarchie; \code{FALSE}, renvoie le premier parent.}

\item{garder_na}{\code{TRUE}, les codes sans parent passent à \code{NA}; \code{FALSE}, les codes sans parent sont gardés tels quels.}
}
\value{
Un vecteur contenant les mentions de diplôme parent.
}
\description{
Renvoie la mention de diplôme parent.
}
