% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchie.R
\name{hier_type_diplome_parent}
\alias{hier_type_diplome_parent}
\title{Renvoie le type de diplome parent}
\usage{
hier_type_diplome_parent(
  code_type_diplome,
  parent_final = FALSE,
  garder_na = FALSE
)
}
\arguments{
\item{parent_final}{\code{TRUE}, renvoit le parent le plus haut dans la hiérarchie; \code{FALSE}, renvoie le premier parent.}

\item{garder_na}{\code{TRUE}, les codes sans parent passent à \code{NA}; \code{FALSE}, les codes sans parent sont gardés tels quels.}

\item{code_mention_diplome}{Un vecteur de codes de type de diplôme.}
}
\value{
Un vecteur contenant les types de diplôme parent.
}
\description{
Renvoie le type de diplôme parent.
}
