% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/historique.R
\name{histo_etape_succ_2}
\alias{histo_etape_succ_2}
\title{Renvoie le code etape successeur (avec prise en charge de l'éclatement)}
\usage{
histo_etape_succ_2(code_etape, successeur_final = TRUE, garder_na = FALSE)
}
\arguments{
\item{code_etape}{Un vecteur de code étape.}

\item{successeur_final}{\code{TRUE}, renvoit le successeur le plus récent dans l'historique; \code{FALSE}, renvoie le premier successeur.}

\item{garder_na}{\code{TRUE}, les codes sans successeur passent à \code{NA}; \code{FALSE}, les codes sans successeur sont gardés tels quels.}
}
\value{
Une liste de code étape successeur.

Jeu de données source : \code{apogee::etape_histo}.\cr
Il est créé à partir d'Apogée et de la table "etape_histo" de la base Access Tables_ref (projet Apogee).
}
\description{
Renvoie le code étape successeur (avec prise en charge de l'éclatement).
}
