% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libelle.R
\name{lib_cycle}
\alias{lib_cycle}
\title{Renvoie le libelle a partir du code cycle de formation}
\usage{
lib_cycle(code_cycle)
}
\arguments{
\item{code_cycle}{Un vecteur de code cycle.}
}
\value{
Un vecteur contenant les libellés de cycle.

Jeu de données source : \code{apogee::cycle}.\cr
}
\description{
Renvoie le libellé à partir du code cycle de formation.
}
\examples{
apogee::lib_cycle(c("1", "3"))

}
