% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libelle.R
\name{lib_etape}
\alias{lib_etape}
\title{Renvoie le libelle a partir du code etape}
\usage{
lib_etape(code_etape, type_diplome = TRUE, annee_etape = TRUE,
  ville = TRUE, option = TRUE, particularite = TRUE)
}
\arguments{
\item{code_etape}{Un vecteur de code étape.}

\item{type_diplome}{\code{TRUE}: type de dplôme intégré dans le libellé d'étape.}

\item{annee_etape}{\code{TRUE}: année d'étape intégrée dans le libellé d'étape.}

\item{ville}{\code{TRUE}: Ville intégrée dans le libellé d'étape.}

\item{option}{\code{TRUE}: option intégrée dans le libellé d'étape.}

\item{particularite}{\code{TRUE}: Particularité intégrée dans le libellé d'étape.}
}
\value{
Un vecteur contenant les libellés étape.

Jeu de données source : \code{apogee::etape}.\cr
Il est créé à partir de la table "etape" de la base Access "Tables_ref.accdb" (projet Apogée).
}
\description{
Renvoie le libellé à partir du code étape.
}
