% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/libelle.R
\name{lib_etape}
\alias{lib_etape}
\title{Renvoie le libelle a partir du code etape}
\usage{
lib_etape(
  code_etape,
  prefixe = "formation",
  suffixe = c("ville", "option", "particularite", "annee")
)
}
\arguments{
\item{code_etape}{Un vecteur de code étape.}

\item{prefixe}{Libellé de type de formation ou de type de diplôme en préfixe du libellé de formation.}

\item{suffixe}{Libellé de ville, option, particularité ou année dans le diplôme en suffixe du libellé de formation.}
}
\value{
Un vecteur contenant les libellés étape.

Jeu de données source : \code{apogee::etape}.\cr
Il est créé à partir d'Apogée et enrichi par la table "etape" de la base Access "Tables_ref.accdb".
}
\description{
Renvoie le libellé à partir du code étape.
}
