% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apogee.R
\name{temoin_etape_actif}
\alias{temoin_etape_actif}
\title{Temoin TRUE/FALSE si le code_etape est actif}
\usage{
temoin_etape_actif(code_etape, annee = NULL, annules = FALSE)
}
\arguments{
\item{code_etape}{Un vecteur de code_etape.}

\item{annee}{Année(s) d'activité à vérifier, par défaut l'année en cours.}

\item{annules}{Si \code{TRUE}, inclut les formations dont toutes les inscriptions sont annulées.}
}
\value{
Un vecteur de booléens TRUE/FALSE.
}
\description{
Témoin TRUE/FALSE si le code_etape est actif.
}
